+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/domain/openid/credentialrequest/holderbinding/AttestationJwt.java
@@ -21,7 +21,7 @@ import java.util.stream.Collectors;
 @Getter
 public final class AttestationJwt {
 
-    @Deprecated(since = "OID4VCI 1.0")
+    @Deprecated(since = "OID4VCI 1.0") // remove later
     private static final String KEY_ATTESTATION_TYPE_ID1 = "keyattestation+jwt";
     private static final Set<AttackPotentialResistance> SUPPORTED_ATTACK_POTENTIAL_RESISTANCE = Set.of(AttackPotentialResistance.ISO_18045_ENHANCED_BASIC, AttackPotentialResistance.ISO_18045_HIGH);
     private static final Set<String> ALLOWED_TYPES = Set.of(KEY_ATTESTATION_TYPE_ID1, "key-attestation+jwt");