+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/service/credential/HolderBindingService.java
@@ -106,7 +106,7 @@ public class HolderBindingService {
         try {
             return credentialRequest.getProofs(
                     applicationProperties.getAcceptableProofTimeWindowSeconds(),
-                    applicationProperties.getAcceptableProofTimeWindowSeconds());
+                    applicationProperties.getNonceLifetimeSeconds());
         } catch (IllegalArgumentException e) {
             throw new Oid4vcException(e, INVALID_CREDENTIAL_REQUEST, "Invalid proof",
                     Map.of("proofType", credentialRequest.getProof() != null ? credentialRequest.getProof().toString() : "null"));