+++ issuer-service/src/test/java/ch/admin/bj/swiyu/issuer/dto/issuance/CredentialRequestDtoTest.java
@@ -1,8 +1,8 @@
-package ch.admin.bj.swiyu.issuer.dto.issuance_v2;
+package ch.admin.bj.swiyu.issuer.dto.issuance;
 
 
-import ch.admin.bj.swiyu.issuer.dto.oid4vci.issuance_v2.CredentialEndpointRequestDtoV2;
-import ch.admin.bj.swiyu.issuer.dto.oid4vci.issuance_v2.ProofsDto;
+import ch.admin.bj.swiyu.issuer.dto.oid4vci.issuance.CreateCredentialRequestDto;
+import ch.admin.bj.swiyu.issuer.dto.oid4vci.issuance.ProofsDto;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
 import jakarta.validation.ValidatorFactory;
@@ -13,13 +13,13 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
-class CredentialRequestDtoV2Test {
+class CredentialRequestDtoTest {
 
     @Test
     void shouldCreateDtoWithAllFields() {
         ProofsDto proofs = new ProofsDto(List.of("jwt"));
 
-        CredentialEndpointRequestDtoV2 dto = new CredentialEndpointRequestDtoV2("config-id", proofs, null);
+        CreateCredentialRequestDto dto = new CreateCredentialRequestDto("config-id", proofs, null);
 
         assertEquals("config-id", dto.credentialConfigurationId());
         assertSame(proofs, dto.proofs());
@@ -30,8 +30,8 @@ class CredentialRequestDtoV2Test {
     void shouldFailValidationIfCredentialConfigurationIdBlank() {
         try (ValidatorFactory factory = Validation.buildDefaultValidatorFactory()) {
             Validator validator = factory.getValidator();
-            var credentialRequestDtoV2 = new CredentialEndpointRequestDtoV2(" ", null, null);
-            var violations = validator.validate(credentialRequestDtoV2);
+            var credentialRequestDto = new CreateCredentialRequestDto(" ", null, null);
+            var violations = validator.validate(credentialRequestDto);
 
             assertThat(violations).isNotEmpty();
             assertThat(violations.stream().map(v -> v.getPropertyPath().toString()).toList()).contains("credentialConfigurationId");