+++ CHANGELOG.md
@@ -13,8 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Breaking Contract change -> Removed c_nonce from OAuthTokenDto the nonce can be retrieved from the nonce endpoint.
-    - The nonce column from credential_offer table is also removed.
+- Breaking Contract change:
+  - Removed c_nonce from OAuthTokenDto the nonce can be retrieved from the nonce endpoint.
+      - The nonce column from credential_offer table is also removed.
+  - Removed Deprecated OID4VCI Draft 13 Endpoints
 
 ### Fixed